Introduce a new S3 repository option "path_style_access"
When using a custom S3 endpoint to an S3-compatible service one typically doesn't want
to use the AWS SDK defaults for using bucket names as sub-domains, but rather the S3 client
should talk to exactly the provided endpoint.

By setting 'path_style_access' to 'true' this can be enforced now, either on the repository
level, or in the default component settings.